MacOS Patch (#624)
* instead of size_t used uint64 when uint64 is expected

* Removed useless files.

* reverted sample folder to default.

+++ src/controller/src/system/rocprofvis_controller_trace_system.cpp
@@ -1727,7 +1727,7 @@ rocprofvis_result_t SystemTrace::AsyncFetch(rocprofvis_property_t property, Futu
                 }
                 case kRPVControllerSystemBucketDataValueIndexed:
                 {
-                    size_t buckets_num = 0;
+                    uint64_t buckets_num = 0;
                     result = GetUInt64(kRPVControllerSystemGetHistogramBucketsNumber, 0, &buckets_num);
                     if (result == kRocProfVisResultSuccess)
                     {
